Extract Vivado code from to a separate file
